feat(map):swipe-tool in component
