+++ packages/integration/src/lib/map/advanced-map-tool/advanced-swipe/advanced-swipe.component.html
@@ -1,0 +1,49 @@
+<div *ngIf="userControlledLayerList.length; else noLayersBlock" class="igo-swipe-select-container">
+    <form class="igo-form" [formGroup]="form">
+        <mat-form-field>
+            <mat-label>{{'igo.integration.advanced-map-tool.swipe-select' | translate}}</mat-label>
+            <mat-select formControlName="layers" multiple>
+                <mat-option [value]="1" (click)="selectAll(e)" #e>
+                    {{'igo.integration.advanced-map-tool.selectAll' | translate}}
+                </mat-option>
+                <mat-divider></mat-divider>
+                <mat-option *ngFor="let layer of userControlledLayerList" [value]="layer" (click)="applyNewsLayers(e)">{{layer.title}}</mat-option>
+            </mat-select>
+        </mat-form-field>
+    </form>
+
+    <mat-slide-toggle class="swipe-toggle"
+        [checked]="swipe"
+        [labelPosition]="'before'"
+        (change)="startSwipe($event.checked)">
+        {{'igo.integration.advanced-map-tool.swipe' | translate}}
+    </mat-slide-toggle> 
+</div>
+
+<ng-template #noLayersBlock>
+    <mat-list>
+        <p class="map-empty mat-typography">
+            {{'igo.integration.advanced-map-tool.empty' | translate}}</p>
+        <p class="map-empty mat-typography">
+            {{'igo.integration.advanced-map-tool.customize' | translate}}</p>
+        <mat-list-item>
+            <mat-icon mat-list-icon svgIcon="magnify"></mat-icon>
+            <h4 matLine class="search-tool mat-typography" (click)="searchEmit()">
+                {{'igo.integration.advanced-map-tool.search-tool' | translate}}
+            </h4>
+        </mat-list-item>
+        <mat-list-item>
+            <mat-icon mat-list-icon svgIcon="layers-plus"></mat-icon>
+            <h4 matLine class="catalog-tool mat-typography" (click)="catalogEmit()">
+                {{'igo.integration.advanced-map-tool.catalog-tool' | translate}}
+            </h4>
+        </mat-list-item>
+        <mat-list-item>
+            <mat-icon mat-list-icon svgIcon="star"></mat-icon>
+            <h4 matLine class="context-tool mat-typography" (click)="contextEmit()">
+                {{'igo.integration.advanced-map-tool.context-tool' | translate}}
+            </h4>
+        </mat-list-item>
+    </mat-list>
+</ng-template>
+
